Updated end to end tests to not call the configdb directly.
